Rename tests/fixtures -> tests/utils
After cleanup, utilities has been split into fixtures in conftest.py and
generic testing utilities (e.g. custom check/assertion functions) in
utils.py.
