+++ nbdime/tests/test_diff.py
@@ -14,7 +14,7 @@ from nbdime import diff
 from nbdime.diff_format import op_patch, op_add, op_replace, op_remove
 from nbdime.diffing.snakes import compute_snakes, compute_snakes_multilevel
 
-from .fixtures import check_symmetric_diff_and_patch
+from .utils import check_symmetric_diff_and_patch
 
 
 def test_diff_and_patch():
