+++ nbdime/tests/utils.py
